Change PyProject to take in the file_path
This also converts the instance variables to be private
and has __load_toml_file convert the file_path to a Path
if it is a str. We also no longer store the file_path
since __load_toml_file is the only place it is needed
so it is just passed in.