TaskRunner only runs the commands now with the project passed in
Instead of putting the commands together, that is now delegated
to Task. The run_commands_and_bail_on_first_fail method is now
gone as that is what `run` does now. Furthermore, the instance
variables of TaskRunner are private now. A single instance of the
TaskRunner can now run many tasks now, since it is not tied to
any particular task.