Refactor taskipy into Task, PyProject, and TaskRunner (#15)
* Run `poetry update`

* Sort test imports & convert unused f strings to strings

There were some f strings that weren't interpolating anything.

* Create custom exceptions

* Refactor taskipi package into task, pyproject, and task_runner

* Add docstrings & types to task & task_runner.

Also, a small refactoring of __bail_on_first_fail()
in task_runner.py.

* Remove docstrings

* Add MissingTaskipySettingsSectionError

* Make path & items in PyProject private

* Check runner is string rather than catching AttributeError

* Add default exception case to cli

* Add exit_code class variable to exceptions

* Change PyProject to take in the file_path

This also converts the instance variables to be private
and has __load_toml_file convert the file_path to a Path
if it is a str. We also no longer store the file_path
since __load_toml_file is the only place it is needed
so it is just passed in.

* Task now builds the commands & takes in args + project

I think args could possibly be moved over to commands
since that is the only place it is used and that would
reduce the amount of things we are passing in to Task.
commands would just have to be converted to a method instead
of a property then.

* TaskRunner only runs the commands now with the project passed in

Instead of putting the commands together, that is now delegated
to Task. The run_commands_and_bail_on_first_fail method is now
gone as that is what `run` does now. Furthermore, the instance
variables of TaskRunner are private now. A single instance of the
TaskRunner can now run many tasks now, since it is not tied to
any particular task.

* Update cli to use updated TaskRunner

The pyproject.toml path is now passed in
with the command name and args being passed
to `run`.

* Use single quotes

* Extract pre_task & post_task to use __find_hooks

They are basically the same method, with a one word
difference.