Add requirements.txt for py36
The latest version of Poetry doesn't support py36.
Furthermore, the lockfile format is different between
the version of poetry supported with py36 and the latest
poetry version. This is a workaround by exporting the
dependencies to a requirements.txt file on CI for py36.

I tried using the lowest common denomenator (the old version
of poetry), but the dependency resolver doesn't seem to be as
good. It never resolves with the current pyproject.toml updates.
However, it does so right away with the updated poetry version.

Since we're planning on removing py36 in the near future, this
tradeoff seems okay in the meantime.