Fix resolution issues on poetry v1.1.5
The resolver on this version isn't quite as good as
newer poetry versions. To help it out, the dev dependencies
are exact versions so it doesn't have to sift through as many
combinations/options. Without this, I was getting into a state
where it was resolving dependencies for over 30,000 seconds and
showed no signs of stopping anytime soon.